// models/Account.ts
/**
 * delivex API
 * Documentation of API endpoints of delivex
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Account {
    'user': number;
    'balance'?: string;
    'availableBalance': string;
    'postpaymentEnabled': string;
    'limit'?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "user",
            "baseName": "user",
            "type": "number",
            "format": ""
        },
        {
            "name": "balance",
            "baseName": "balance",
            "type": "string",
            "format": "decimal"
        },
        {
            "name": "availableBalance",
            "baseName": "available_balance",
            "type": "string",
            "format": ""
        },
        {
            "name": "postpaymentEnabled",
            "baseName": "postpayment_enabled",
            "type": "string",
            "format": ""
        },
        {
            "name": "limit",
            "baseName": "limit",
            "type": "string",
            "format": "decimal"
        }    ];

    static getAttributeTypeMap() {
        return Account.attributeTypeMap;
    }

    public constructor() {
    }
}