// models/AccountTransaction.ts
/**
 * delivex API
 * Documentation of API endpoints of delivex
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { AccountTransactionTypeEnum } from '../models/AccountTransactionTypeEnum';
import { TransactionOrder } from '../models/TransactionOrder';
import { HttpFile } from '../http/http';

export class AccountTransaction {
    'order'?: TransactionOrder | null;
    'amount': string;
    'type': AccountTransactionTypeEnum;
    'createdAt': Date;
    'accountBalanceSnapshot': string;
    'typeDisplayName': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "order",
            "baseName": "order",
            "type": "TransactionOrder",
            "format": ""
        },
        {
            "name": "amount",
            "baseName": "amount",
            "type": "string",
            "format": "decimal"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "AccountTransactionTypeEnum",
            "format": ""
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "accountBalanceSnapshot",
            "baseName": "account_balance_snapshot",
            "type": "string",
            "format": "decimal"
        },
        {
            "name": "typeDisplayName",
            "baseName": "type_display_name",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AccountTransaction.attributeTypeMap;
    }

    public constructor() {
    }
}