// models/Address.ts
/**
 * delivex API
 * Documentation of API endpoints of delivex
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class Address {
    'street': string;
    'streetNumber': string;
    'apartmentNumber'?: string;
    'city': string;
    'region'?: string;
    'country'?: string;
    'longitude': string;
    'latitude': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "street",
            "baseName": "street",
            "type": "string",
            "format": ""
        },
        {
            "name": "streetNumber",
            "baseName": "street_number",
            "type": "string",
            "format": ""
        },
        {
            "name": "apartmentNumber",
            "baseName": "apartment_number",
            "type": "string",
            "format": ""
        },
        {
            "name": "city",
            "baseName": "city",
            "type": "string",
            "format": ""
        },
        {
            "name": "region",
            "baseName": "region",
            "type": "string",
            "format": ""
        },
        {
            "name": "country",
            "baseName": "country",
            "type": "string",
            "format": ""
        },
        {
            "name": "longitude",
            "baseName": "longitude",
            "type": "string",
            "format": "decimal"
        },
        {
            "name": "latitude",
            "baseName": "latitude",
            "type": "string",
            "format": "decimal"
        }    ];

    static getAttributeTypeMap() {
        return Address.attributeTypeMap;
    }

    public constructor() {
    }
}