// models/CancelOrderRequest.ts
/**
 * delivex API
 * Documentation of API endpoints of delivex
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CancelOrderRequest {
    'cancellationWithPayment'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "cancellationWithPayment",
            "baseName": "cancellation_with_payment",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CancelOrderRequest.attributeTypeMap;
    }

    public constructor() {
    }
}