// models/City.ts
/**
 * delivex API
 * Documentation of API endpoints of delivex
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class City {
    'id': number;
    'name': string;
    'intracityDeliveryEnabled'?: boolean;
    'longitude'?: string | null;
    'latitude'?: string | null;
    'region'?: number | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "intracityDeliveryEnabled",
            "baseName": "intracity_delivery_enabled",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "longitude",
            "baseName": "longitude",
            "type": "string",
            "format": "decimal"
        },
        {
            "name": "latitude",
            "baseName": "latitude",
            "type": "string",
            "format": "decimal"
        },
        {
            "name": "region",
            "baseName": "region",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return City.attributeTypeMap;
    }

    public constructor() {
    }
}